requirements_fixed.txt
# Federated Learning
flwr>=1.6.0,<1.8.0

# Machine Learning & Deep Learning (using stable versions)
tensorflow>=2.13.0,<2.16.0
torch>=2.0.0,<2.3.0
scikit-learn>=1.3.0,<1.6.0
opencv-python>=4.8.0,<4.10.0

# Data Handling
pandas>=2.0.0,<2.3.0
numpy>=1.24.0,<1.27.0

# Web3 & API
web3>=6.0.0,<7.0.0
fastapi>=0.100.0,<0.112.0
uvicorn>=0.20.0,<0.30.0
aiohttp>=3.8.0,<4.0.0

# Database & Environment
SQLAlchemy>=2.0.0,<2.1.0
python-dotenv>=1.0.0,<1.1.0

# Visualization & Dashboard
streamlit>=1.30.0,<1.40.0
plotly>=5.15.0,<5.23.0
folium>=0.15.0,<0.18.0

# Utilities
psutil>=5.9.0,<6.0.0
cryptography>=41.0.2,<42.0.0
pytest>=7.0.0,<9.0.0

# Additional dependencies to fix grpc issues
grpcio>=1.50.0,<1.60.0
protobuf>=3.19.0,<5.0.0